+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Polyverse Boost Cloud Service
 
 # Bug Fixes
 - Fixed token max calculation when using non-GPT4 models
+- Fix regression in custom processor for custom prompts
+- Fix regression in blueprint processor for custom seeds
+- Fix issue with custom processor sending too many messages - truncate more than 60% of total token buffer
+- Fix issue with missing content in custom processor messages
 
 ## Version 0.6.1: June 19, 2023